refactor: 404 페이지 서버 페이지로 변경 - 서버 페이지로 변경 - 버튼 컴포넌트로 분리
